requirements.txt
pyparsing==2.0.3
Django==1.8.2
django-blog-zinnia==0.15.1
django-bootstrap3==5.1.1
django-mptt==0.7.3
django-suit==0.2.12
django-ckeditor==4.4.8
zinnia-wysiwyg-ckeditor==1.2
django-settings-export==1.0.3
-e git://github.com/eugena/django-secretballot.git#egg=django-secretballot

